+++ src/BoxKite.Twitter/Models/Connection/TwitterAccountsDictionary.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 namespace BoxKite.Twitter.Models
 {
-    public class TwitterAccountsDictionary : Dictionary<int,TwitterAccount>
+    public class TwitterAccountsDictionary : Dictionary<long,TwitterAccount>
     {
     }
 }
